Fixed behat tests
