+++ features/backend/countries.feature
@@ -6,12 +6,12 @@ Feature: Countries and provinces
 
     Background:
         Given there is default currency configured
-        And there is default channel configured
         And there are following locales configured:
             | code  | enabled |
             | en_US | yes     |
-          And I am logged in as administrator
-          And there are following countries:
+        And there is default channel configured
+        And I am logged in as administrator
+        And there are following countries:
             | name    | provinces                       |
             | France  | Lyon, Toulouse, Rennes, Nancy   |
             | China   |                                 |
