+++ features/backend/taxonomies.feature
@@ -6,10 +6,10 @@ Feature: taxonomies
 
     Background:
         Given there is default currency configured
-        And there is default channel configured
         And there are following locales configured:
             | code  | enabled |
             | en_US | yes     |
+          And there is default channel configured
           And I am logged in as administrator
           And there are following taxonomies defined:
             | name     |
