+++ features/frontend/cart.feature
@@ -22,12 +22,15 @@ Feature: Cart
             | Git T-Shirt   | 29.99 | T-Shirt size                | PHP T-Shirts | match              |
             | PHP Top       | 5.99  |                             | PHP T-Shirts |                    |
             | iShirt        | 18.99 |                             | T-Shirts     |                    |
-          And all products assigned to "DEFAULT-WEB" channel
           And product "Super T-Shirt" is available in all variations
           And product "Git T-Shirt" is available in all variations
           And product "Black T-Shirt" is available in all variations
           And there is default currency configured
           And there is default channel configured
+          And channel "DEFAULT-WEB" has following configuration:
+              | taxonomy |
+              | Category |
+          And all products assigned to "DEFAULT-WEB" channel
 
     Scenario: Seeing empty cart
         Given I am on the store homepage
