+++ features/frontend/cart_inclusive_tax.feature
@@ -25,6 +25,9 @@ Feature: Tax included in price
             | name    | price | taxons       | tax category  |
             | PHP Top | 85    | PHP T-Shirts | Taxable Goods |
           And all products assigned to "DEFAULT-WEB" channel
+          And channel "DEFAULT-WEB" has following configuration:
+            | taxonomy |
+            | Category |
           And the default tax zone is "Germany"
 
     Scenario: Correct amounts are displayed for inclusive taxes
