+++ features/frontend/cart_promotions_coupons.feature
@@ -43,6 +43,7 @@ Feature: Checkout coupon promotions
         And there is default currency configured
         And there is default channel configured
         And all products assigned to "DEFAULT-WEB" channel
+        And all promotions assigned to "DEFAULT-WEB" channel
 
     Scenario: Promotion with coupons is applied when the customer
               has added a valid coupon
