+++ features/frontend/cart_promotions_dates.feature
@@ -35,6 +35,7 @@ Feature: Checkout limited time promotions
         And there is default currency configured
         And there is default channel configured
         And all products assigned to "DEFAULT-WEB" channel
+        And all promotions assigned to "DEFAULT-WEB" channel
 
     Scenario: Promotion is applied when the order date corresponds
               with promotion dates
