+++ features/frontend/cart_taxation.feature
@@ -27,6 +27,10 @@ Feature: Cart taxation
           And there is default currency configured
           And there is default channel configured
           And all products assigned to "DEFAULT-WEB" channel
+          And channel "DEFAULT-WEB" has following configuration:
+            | taxonomy |
+            | Category |
+
 
     Scenario: No taxes are applied for unknown billing address
               when default tax zone is not configured
