+++ features/frontend/checkout_payment.feature
@@ -27,6 +27,9 @@ Feature: Checkout Payment
             | PayPal      | paypal     | yes     |
             | PayPal PRO  | paypal_pro | no      |
           And all products assigned to "DEFAULT-WEB" channel
+          And channel "DEFAULT-WEB" has following configuration:
+            | taxonomy | payment                           | shipping    |
+            | Category | PayPal, PayPal PRO, Credit Card   | DHL Express |
           And I am logged in user
           And I added product "PHP Top" to cart
           And I go to the checkout start page
