+++ features/frontend/checkout_taxation.feature
@@ -31,6 +31,9 @@ Feature: Checkout taxation
           And there is default currency configured
           And there is default channel configured
           And all products assigned to "DEFAULT-WEB" channel
+          And channel "DEFAULT-WEB" has following configuration:
+            | taxonomy | payment | shipping    |
+            | Category | Dummy   | DHL Express |
           And I am logged in user
           And I added product "PHP Top" to cart
           And I go to the checkout start page
