+++ src/Sylius/Bundle/CoreBundle/spec/Cart/ItemResolverSpec.php
@@ -88,7 +88,7 @@ class ItemResolverSpec extends ObjectBehavior
         $request->isMethod('POST')->willReturn(true);
         $request->get('id')->willReturn(5);
 
-        $productRepository->findOneBy(array('id' => 5, 'channel' => null))->willReturn(null);
+        $productRepository->findOneBy(array('id' => 5, 'channels' => null))->willReturn(null);
 
         $this
             ->shouldThrow('Sylius\Component\Cart\Resolver\ItemResolvingException')
