+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Order/macros.html.twig
@@ -27,7 +27,7 @@
             <td class="center-text"><input type="checkbox" value="{{ order.id }}" /></td>
             <td>{{ order.createdAt|date }}</td>
             <td>
-                <span style="color: {{ order.channel.color }};">{{ order.channel.code }}</span>
+                <span {%- if order.channel.color -%} style="color: {{ order.channel.color }};"{%- endif -%}>{{ order.channel.code }}</span>
             </td>
             <td>
                 <a href="{{ path('sylius_backend_order_show', {'id': order.id}) }}">
