+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Payment/macros.html.twig
@@ -24,7 +24,7 @@
 
         <tr id="{{ payment.id }}">
             <td>
-                <a href="{{ path('sylius_backend_order_show', {'id': payment.order.id}) }}" class="btn btn-primary btn-sm">
+                <a href="{{ path('sylius_backend_order_show', {'id': order.id}) }}" class="btn btn-primary btn-sm">
                    #{{ order.number }}
                 </a>
             </td>
